déplacement des autres fonctions de publication
